+++ src/Service/Athena/src/AthenaClient.php
@@ -729,6 +729,7 @@ class AthenaClient extends AbstractApi
             case 'eu-west-1':
             case 'eu-west-2':
             case 'eu-west-3':
+            case 'il-central-1':
             case 'me-central-1':
             case 'me-south-1':
             case 'sa-east-1':